fix: resolved all issues, finish PR
